server/requirements.txt
fastapi
uvicorn[standard]
python-multipart

#langchain and its dependencies
langchain
langchain-community
langchain-core
langchain-groq

pinecone


langchain-google-genai

pyPDF

python-dotenv

pydantic
requests
tqdm

loguru

aiofiles

streamlit

